tools/ModelMeta/requirements.txt
absl-py==0.15.0
addict==2.4.0
appdirs==1.4.4
array_record==0.5.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
autograd==1.7.0
autoray==0.6.12
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
chex==0.1.86
clang==5.0
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
dill==0.3.8
dm-tree==0.1.8
easydict==1.13
et-xmlfile==1.1.0
etils==1.5.2
exceptiongroup==1.2.1
executing==2.0.1
faiss-gpu==1.7.2
filelock==3.14.0
flatbuffers==1.12
flax==0.8.4
fonttools==4.53.0
fsspec==2024.5.0
ftfy==6.2.0
future==1.0.0
gast==0.4.0
gensim==4.3.2
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.64.0
h5py==3.1.0
huggingface-hub==0.23.2
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipython==8.18.1
jax==0.4.25
jaxlib==0.4.25+cuda11.cudnn86
jedi==0.19.1
jieba==0.42.1
joblib==1.4.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
latex2mathml==3.77.0
lazy_loader==0.4
Levenshtein==0.25.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.0
matplotlib-inline==0.1.7
mdtex2html==1.3.0
mdurl==0.1.2
mindformers==1.1.0
mindpet==1.0.4
mindspore @ file:///home/cvgroup/myz/czx/3.9/mindspore-2.2.14-cp39-cp39-linux_x86_64.whl#sha256=66d1864ccb722c82c1fa176b6277d33301b700b326e4932a7fa30209ccce57c1
ml-dtypes==0.4.0
mmcv==2.1.0
mmengine==0.10.4
msgpack==1.0.8
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.22.0
oauthlib==3.2.2
onnx==1.10.1
onnxruntime==1.9.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.5
opt-einsum==3.3.0
optax==0.2.2
orbax-checkpoint==0.5.14
packaging==24.0
pandas==2.0.0
parso==0.8.4
PennyLane==0.38.0
PennyLane_Lightning==0.38.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
prettytable==3.7.0
promise==2.3
prompt_toolkit==3.0.45
protobuf==3.20.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pyDeprecate==0.3.2
Pygments==2.18.0
pynvml==11.5.0
pyparsing==3.1.2
PySnooper==1.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.3
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rouge-chinese==1.0.3
rsa==4.9
rustworkx==0.15.1
scikit-image==0.22.0
scikit-learn==1.5.0
scipy==1.10.1
seaborn==0.13.2
sentencepiece==0.2.0
six==1.15.0
smart-open==7.0.4
stack-data==0.6.3
tdqm==0.0.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.0
tensorflow-datasets==2.1.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-metadata==1.15.0
tensorstore==0.1.60
termcolor==1.1.0
threadpoolctl==3.5.0
tifffile==2024.5.22
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch @ file:///home/cvgroup/myz/czx/3.9/torch-1.13.1%2Bcu117-cp39-cp39-linux_x86_64.whl#sha256=b3ac139e0d4a0b303cc16f51eb77146ec7d14c0b1e702ad638613628f5086af7
torchaudio @ file:///home/cvgroup/myz/czx/3.9/torchaudio-0.13.1%2Bcu117-cp39-cp39-linux_x86_64.whl#sha256=65bb3615d433d5b930acdc1036efb124047d5717dba4deb70f419290d942f836
torchmetrics==0.7.0
torchsummary==1.5.1
torchvision @ file:///home/cvgroup/myz/czx/3.9/torchvision-0.14.1%2Bcu117-cp39-cp39-linux_x86_64.whl#sha256=8a89a0ee007d7cdba53bd56426d8795060a064b8919b61ac39e00e399ded6be2
tqdm==4.66.4
traitlets==5.14.3
troubleshooter==1.0.17
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.12.1
yacs==0.1.8
yapf==0.40.2
zipp==3.19.0
